requirements.txt
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
streamlit==1.33.0
plotly==5.20.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.29
dbt-postgres==1.8.0
